Decided to alter behavior of packetpeek slightly:
packetpeek can report -1 indicating a hole in the data (not really an
error).

Previously, packetout would have to be called to clear the hole; now,
the hole is cleared by packetpeek.  That is, the change is that
ogg_stream_packetpeek must be followed by a call to
ogg_stream_packetout(state,NULL) to continue advancing forward only if
packetpeek returned a packet.

svn path=/trunk/ogg/; revision=1454