Explicitly document what ogg_packet_clear() does.
The ogg_packet structure is public, and contains no 'internal' data,
except in that decoding functions allocate and own the returned buffer.
When a client creates packets to pass into the encoding functions it
is instead the client which owns the buffer. Nevertheless,
ogg_packet_clear() can be useful if the client is using a simple
allocation strategy. I wanted to be clear that it was appropriate
to use this call for such packets.

svn path=/trunk/ogg/; revision=17497