Add gitlab integration tests.
We've moved hosting to a gitlab instance. Make use of the built-in
integration testing framework to verify commits.

This has parallel jobs to run the autotools and cmake builds.

Note that the autotools job needs cmake because the `distcheck`
target verifies that the generated source package also builds
correctly with cmake, in particular to check that the necessary
files are included.