Don't update homebrew on macOS travis builds.
The macOS system image already provides a version of xz through
homebrew, so we don't need to explicitly install it through
`brew bundle`.

Skipping the `brew update` step reduces the runtime of test jobs
from 6 minutes to 1, so we get test feedback more promptly.

The xz package is really only needed for 'make dist' targets,
so listing it in a Brewfile doesn't help users get started
building the library. Better not to clutter the top-level
directory further.