github actions: Add a GNU autotools build.
This only runs on linux and macOS, but supports the important
`distcheck` target.

We ask homebrew to install libtool and pkg-config on macOS.
These are required for the autotools build, but aclocal
at least seems not to be available in the latest image.

We don't do a general homebrew update since that can take several
minutes, at least doubling the time it takes to runthhe job.
Just using the defaults the image provides is a better tradeoff.